// src/dex/algebra/types.ts
export type GlobalStateV1_1 = {
  price: bigint; // The square root of the current price in Q64.96 format
  tick: bigint; // The current tick
  fee: bigint; // The current fee in hundredths of a bip, i.e. 1e-6
  communityFeeToken0: bigint; // The community fee represented as a percent of all collected fee in thousandths (1e-3)
  communityFeeToken1: bigint;
};

export type PoolStateV1_1 = {
  pool: string;
  blockTimestamp: bigint;
  tickSpacing: bigint; // is actually constant
  globalState: GlobalStateV1_1; // eq slot0
  liquidity: bigint;
  maxLiquidityPerTick: bigint; // is actually constant
  tickBitmap: Record<NumberAsString, bigint>; // actually called tickTable in contract-
  ticks: Record<NumberAsString, TickInfo>; // although variable names are different in contracts but matches UniswapV3 TickInfo struct 1:1
  isValid: boolean;
  startTickBitmap: bigint;
  balance0: bigint;
  balance1: bigint;
  areTicksCompressed: boolean;
};